modulo-05-compose/projeto-taskmanager/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.7
